add a script which extracts metadata from a pdf given on command line
